+++ lib/app.js
@@ -1,5 +1,5 @@
-const express = require('express');
 const path = require('path');
+const express = require('express');
 const app = express();
 
 const sander = require('sander');
@@ -13,32 +13,83 @@ app.get('/favicon.ico', (req, res) => {
   favicon.pipe(res);
 });
 
+const filePath = path.join(__dirname, '/store');
 
-function bodyParser(req, res) {
+app.get('/store', function(req, res) {
 
-  var body = '';
+  let files = '';
+  let body;
+
+  sander.readdir(filePath)
+    .then((data) => {
+      data.forEach((el) => {
+        files += el + '\n';
+      });
+      body = {files};
+      res.send(body);
+      res.end();
+    })
+    .catch((err) => {
+      console.log(err);
+    });
+
+});
+
+app.get('/store/:fileName', function(req, res) {
+
+  let file = '';
+  let body;
+
+  let fileName = req.params.fileName;
+  let fullPath = path.join(filePath, fileName); 
+
+  sander.readFile(fullPath)
+    .then((data) => {
+      file += data.toString('utf8');
+      file = JSON.parse(file);
+      body = file;
+      res.send(body);
+      res.end();
+    });
+
+});
+
+app.post('/store', function(req, res) {
+  
+  let body = '';
 
   req.on('data', (chunk) => {
     body += chunk;
   });
 
   req.on('end', () => {
-    if(body) {
-      req.body = JSON.parse(body);
-    }
-    onEnd(req, res);
+    req.body = JSON.parse(body);
+
+    let file;
+    
+    let book = JSON.stringify(req.body);
+
+    let bookName = req.body.title;
+    let fileName = bookName.replace(/\s/g, '_');
+
+    file = path.join(filePath, fileName + '.json');
+
+    fs.writeFile(file, book, (err) => {
+
+      if(err) {
+        throw err;
+      } else {
+        body = {fileName: fileName};
+        res.send(body);
+        res.end();
+      }
+
+    });
   });
-
-}
-
-const filePath = path.join(__dirname, '/store');
-
-app.get('/store', router);
-
-
+});
 
 app.use('/store', (err, req, res, next) => {
-  const code = error.code || 500;
+  const code = err.code || 500;
   const error = code === 500 ? 'Internal Server Error' : err.error;
   console.log(err.error || err.message);
   res.status(code).send({error});
