+++ lib/store/Life_on_the_Mississippi.json
@@ -1,0 +1,1 @@
+{"title":"Life on the Mississippi","author":"Mark Twain"}
